+++ pychiver/timeutils.py
@@ -7,14 +7,14 @@ import datetime
 VALID_DATE_FORMAT = "%Y-%m-%d %H:%M:%S"
 
 
-def validateTimeStamps(start_date, end_date=None) -> tuple:
+def validateTimeStampsReturnObjects(start_date, end_date=None) -> tuple:
     """
     Ensures that provided time stamps are of correct format.
     Accepted objects: string (with format ) or datetime object
 
     :param start_date:
     :param end_date: default is None, that translates into datetime.now()
-    :return: two formatted strings for start and end date
+    :return: two datetime objects for the start and the end date
 
     :raises ValueError if wrong type of objects provided
     """
@@ -31,7 +31,30 @@ def validateTimeStamps(start_date, end_date=None) -> tuple:
     if not isinstance(end_date, datetime.datetime) and not isinstance(end_date, str):
         raise ValueError('Wrong end_date format (neither date time nor string)!')
 
-    return getTimeStampFormatted(start_date), getTimeStampFormatted(end_date)
+    return getDateTimeObj(start_date), getDateTimeObj(end_date)
+
+
+def validateTimeStamps(start_date, end_date=None) -> tuple:
+    """
+    Ensures that provided time stamps are of correct format.
+    Accepted objects: string (with format ) or datetime object
+
+    :param start_date:
+    :param end_date: default is None, that translates into datetime.now()
+    :return: two formatted strings for the start and the end date
+
+    :raises ValueError if wrong type of objects provided
+    """
+    s, e = validateTimeStampsReturnObjects(start_date, end_date)
+    return getTimeStampFormatted(s), getTimeStampFormatted(e)
+
+
+def getDateTimeObj(date_obj, date_input_format=VALID_DATE_FORMAT) -> datetime.datetime:
+    if isinstance(date_obj, datetime.datetime):
+        pass
+    else:
+        date_obj = datetime.datetime.strptime(date_obj, date_input_format)
+    return date_obj
 
 
 def getTimeStampFormatted(date_obj, date_input_format=VALID_DATE_FORMAT) -> str:
@@ -43,8 +66,5 @@ def getTimeStampFormatted(date_obj, date_input_format=VALID_DATE_FORMAT) -> str:
     :param date_input_format: default "%Y-%m-%d %H:%M:%S"
     :return: ESS Archiver formatted date string
     """
-    if isinstance(date_obj, datetime.datetime):
-        pass
-    else:
-        date_obj = datetime.datetime.strptime(date_obj, date_input_format)
+    date_obj = getDateTimeObj(date_obj, date_input_format)
     return date_obj.isoformat() + "Z"