+++ tests/test_calculations.py
@@ -6,9 +6,11 @@ SIMPLE_VALUES = [i * i for i in SIMPLE_TIMES]
 
 NEW_TIME_BASE_CORRECT = [1, 3, 4.5, 7.66, 10]
 NEW_VALUES_CORRECT_TIME_BASE = [1, 9., 20.5, 58.9, 100.]
+NEW_VALUES_CORRECT_TIME_BASE_LAST_ACQ = [1, 9, 16, 49, 100]
 
 NEW_TIME_BASE_INCORRECT = [-1, 3, 4.5, 7.66, 14]
 NEW_VALUES_INCORRECT_TIME_BASE = [1., 9., 20.5, 58.9, 100.]
+NEW_VALUES_INCORRECT_TIME_BASE_LAST_ACQ = [1, 9, 16, 49, 100]
 
 DF_COLUMNS = ['time', 'val', 'some_other']
 DF_1 = pd.DataFrame(np.array([[1.0, 2, 3],
@@ -71,6 +73,27 @@ class TestLinearInterpolation(unittest.TestCase):
         for i in range(len(array1)):
             self.assertAlmostEqual(array1[i], array2[i], places=2)
 
+
+class TestLastValueInterpolation(unittest.TestCase):
+
+    def test_wrong_init(self):
+        with self.assertRaises(ValueError):
+            lis = LastAcquiredValueInterpolationStrategy(baseXs=[1])
+
+    def test_the_correct_base(self):
+        lis = LastAcquiredValueInterpolationStrategy(baseXs=NEW_TIME_BASE_CORRECT)
+        result = lis.getValues(SIMPLE_TIMES, SIMPLE_VALUES)
+        self._compare_two_arrays(result, NEW_VALUES_CORRECT_TIME_BASE_LAST_ACQ)
+
+    def test_with_the_incorrect_base(self):
+        lis = LastAcquiredValueInterpolationStrategy(baseXs=NEW_TIME_BASE_INCORRECT)
+        result = lis.getValues(SIMPLE_TIMES, SIMPLE_VALUES)
+        self._compare_two_arrays(result, NEW_VALUES_INCORRECT_TIME_BASE_LAST_ACQ)
+
+    def _compare_two_arrays(self, array1, array2):
+        self.assertEqual(len(array1), len(array2))
+        for i in range(len(array1)):
+            self.assertAlmostEqual(array1[i], array2[i], places=2)
 
 class TestDataAlign(unittest.TestCase):
     def test_wrong_init_no_interpolation(self):